blockchain: move static methods to external functions
