+++ packages/blockchain/src/index.ts
@@ -7,4 +7,5 @@ export {
   DBSetHashToNumber,
   DBSetTD,
 } from './db/helpers.js'
+export * from './helpers.js'
 export * from './types.js'
